feat: Add secret name trigger to experience
Introduces a hidden trigger mechanism in the 3D experience that reveals a secret message. This is activated when the user positions the camera behind the pearl at a close distance. The message displays briefly and has a cooldown period before it can be triggered again.

Adds CSS for fade-in-up animation and a new React component for the secret trigger logic. Also modifies the HTML to include the CSS.